Add vote precondition

+++ src/main/java/ru/pkg/web/voting/AbstractVotingController.java
@@ -5,6 +5,8 @@ import ru.pkg.LoggedUser;
 import ru.pkg.model.UserVote;
 import ru.pkg.service.VotingService;
 import ru.pkg.to.VotingStatistics;
+import ru.pkg.utils.TimeUtil;
+import ru.pkg.utils.exception.VotingException;
 
 import java.util.List;
 
@@ -15,13 +17,18 @@ public abstract class AbstractVotingController {
 
     public void vote(int restaurantId) {
         UserVote userVote = service.findById(LoggedUser.getId());
-        if (!userVote.isVotedToday()) {
-            userVote = service.save(LoggedUser.getId(), restaurantId);
-            LoggedUser.getUserTO().setLastVoted(userVote.getLastVoted());
+        if (!TimeUtil.isCanVote() || userVote.isVotedToday()) {
+            throw new VotingException(TimeUtil.isCanVote() ? "You have already voted today." : "Voting finished today.");
         }
+
+        userVote = service.save(LoggedUser.getId(), restaurantId);
+        LoggedUser.getUserTO().setLastVoted(userVote.getLastVoted());
     }
 
     public void cancel() {
+        if (!TimeUtil.isCanVote()) {
+            throw new VotingException("Voting finished today.");
+        }
         service.delete(LoggedUser.getId());
         LoggedUser.getUserTO().setLastVoted(null);
     }
